fix stop script

+++ scripts/stopServer.sh
@@ -1,4 +1,12 @@
 #!/bin/bash
 source /home/ubuntu/.profile
 cd /home/ubuntu/acdsampleapp
-killall node
+
+x=`ps -ef | grep node | grep -v grep | awk '{print $1}'`
+echo $x
+if [ -z "$x" ]
+then
+      echo "No node process found!"
+else
+      killall node
+fi
